requirements.txt
thriftpy
google-cloud-pubsub
sqlalchemy
mysqlclient
pytesseract
pillow
imutils
livescore
pytube
livestreamer
